prepare schema refactor
